# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DevLens API
  version: 1.0.0
  description: |
    API contract สำหรับ DevLens Engineering Intelligence Platform

    Related documents:
    - 01-business-logic-architecture.md
    - 03-backend-roadmap.md
    - 04-metric-definitions.md
    - 05-database-design.dbml
    - 06-api-design.md
servers:
  - url: /api/v1
    description: Current API version
...
security:
  - bearerAuth: []
paths:
...
  /organizations/{organizationId}/insights:
    get:
      tags:
        - Insights
      summary: List generated insights for an organization
      operationId: listInsights
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/RepositoryIdQuery"
        - $ref: "#/components/parameters/InsightType"
        - $ref: "#/components/parameters/InsightStatus"
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Insight list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /insights:
    get:
      tags:
        - Insights
      summary: List generated insights across a selected organization
      operationId: listInsightsAlias
      parameters:
        - $ref: "#/components/parameters/OrganizationIdQuery"
        - $ref: "#/components/parameters/RepositoryIdQuery"
        - $ref: "#/components/parameters/InsightType"
        - $ref: "#/components/parameters/InsightStatus"
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Insight list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /organizations/{organizationId}/insights/{insightKey}/review:
    post:
      tags:
        - Insights
      summary: Mark an insight as reviewed
      operationId: reviewInsight
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/InsightKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInsightRequest"
      responses:
        "200":
          description: Insight status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /organizations/{organizationId}/insights/{insightKey}/dismiss:
    post:
      tags:
        - Insights
      summary: Dismiss an insight
      operationId: dismissInsight
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/InsightKey"
      responses:
        "200":
          description: Insight status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /organizations/{organizationId}/insights/{insightKey}/reopen:
    post:
      tags:
        - Insights
      summary: Reopen an insight
      operationId: reopenInsight
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/InsightKey"
      responses:
        "200":
          description: Insight status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    OrganizationIdQuery:
      name: organizationId
      in: query
      required: true
      schema:
        type: string
        format: uuid
...
    RepositoryIdQuery:
      name: repositoryId
      in: query
      schema:
        type: string
        format: uuid
...
    InsightKey:
      name: insightKey
      in: path
      required: true
      schema:
        type: string
    InsightType:
      name: type
      in: query
      schema:
        type: string
        enum:
          - bottleneck_detection
          - large_pr_detection
          - slow_review_detection
          - hotspot_detection
          - deployment_failure_trend
          - review_concentration
    InsightStatus:
      name: status
      in: query
      schema:
        type: string
        enum:
          - open
          - reviewed
          - dismissed
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    DateFrom:
      name: from
      in: query
      required: true
      description: Inclusive start date
      schema:
        type: string
        format: date
    DateTo:
      name: to
      in: query
      required: true
      description: Inclusive end date
      schema:
        type: string
        format: date
...
  responses:
    Unauthorized:
      description: Authentication is required or token is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: User does not have permission to access the resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
...
    PaginationMeta:
      type: object
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
...
    ReviewInsightRequest:
      type: object
      properties:
        reviewedBy:
          type: string
          format: uuid
    Insight:
      type: object
      required:
        - insightKey
        - insightType
        - status
        - severity
        - title
        - summary
        - organizationId
        - detectedAt
      properties:
        insightKey:
          type: string
        insightType:
          type: string
          enum:
            - bottleneck_detection
            - large_pr_detection
            - slow_review_detection
            - hotspot_detection
            - deployment_failure_trend
            - review_concentration
        status:
          type: string
          enum:
            - open
            - reviewed
            - dismissed
        severity:
          type: string
          enum:
            - low
            - medium
            - high
            - critical
        title:
          type: string
        summary:
          type: string
        organizationId:
          type: string
          format: uuid
        repositoryId:
          type: string
          format: uuid
          nullable: true
        repositoryName:
          type: string
          nullable: true
        detectedAt:
          type: string
          format: date-time
        evidence:
          type: object
          additionalProperties: true
        reviewedBy:
          type: string
          format: uuid
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        dismissedAt:
          type: string
          format: date-time
          nullable: true
        reopenedAt:
          type: string
          format: date-time
          nullable: true
    InsightListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Insight"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
    InsightStatusPayload:
      type: object
      required:
        - insightKey
        - insightType
        - status
        - updatedAt
      properties:
        insightKey:
          type: string
        insightType:
          type: string
        status:
          type: string
          enum:
            - open
            - reviewed
            - dismissed
        reviewedBy:
          type: string
          format: uuid
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        dismissedAt:
          type: string
          format: date-time
          nullable: true
        reopenedAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
    InsightStatusResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/InsightStatusPayload"
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: REPOSITORY_NOT_FOUND
        message:
          type: string
          example: Repository not found
        requestId:
          type: string
          example: req_01JABCXYZ
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ValidationIssue:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: from
        message:
          type: string
          example: from must be before or equal to to
    ValidationErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          allOf:
            - $ref: "#/components/schemas/ErrorDetail"
            - type: object
              properties:
                details:
                  type: array
                  items:
                    $ref: "#/components/schemas/ValidationIssue"